+++ core/src/commonTest/kotlin/com/powersync/sync/SyncStreamTest.kt
@@ -11,6 +11,8 @@ import com.powersync.bucket.BucketStorage
 import com.powersync.bucket.Checkpoint
 import com.powersync.bucket.OpType
 import com.powersync.bucket.OplogEntry
+import com.powersync.bucket.WriteCheckpointData
+import com.powersync.bucket.WriteCheckpointResponse
 import com.powersync.connectors.PowerSyncBackendConnector
 import com.powersync.connectors.PowerSyncCredentials
 import com.powersync.db.crud.CrudEntry
@@ -28,7 +30,10 @@ import dev.mokkery.verify.VerifyMode.Companion.order
 import dev.mokkery.verifyNoMoreCalls
 import dev.mokkery.verifySuspend
 import io.ktor.client.engine.mock.MockEngine
+import kotlinx.coroutines.DelicateCoroutinesApi
+import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.runTest
@@ -105,6 +110,7 @@ class SyncStreamTest {
                     uploadCrud = {},
                     logger = logger,
                     params = JsonObject(emptyMap()),
+                    scope = this,
                 )
 
             syncStream.invalidateCredentials()
@@ -141,10 +147,11 @@ class SyncStreamTest {
                     retryDelayMs = 10,
                     logger = logger,
                     params = JsonObject(emptyMap()),
+                    scope = this,
                 )
 
             syncStream.status.update(connected = true)
-            syncStream.triggerCrudUpload()
+            syncStream.triggerCrudUploadAsync().join()
 
             testLogWriter.assertCount(2)
 
@@ -180,6 +187,7 @@ class SyncStreamTest {
                     retryDelayMs = 10,
                     logger = logger,
                     params = JsonObject(emptyMap()),
+                    scope = this,
                 )
 
             // Launch streaming sync in a coroutine that we'll cancel after verification
@@ -209,7 +217,7 @@ class SyncStreamTest {
             // TODO: It would be neat if we could use in-memory sqlite instances instead of mocking everything
             // Revisit https://github.com/powersync-ja/powersync-kotlin/pull/117/files at some point
             val syncLines = Channel<SyncLine>()
-            val client = MockSyncService(syncLines)
+            val client = MockSyncService(syncLines, { WriteCheckpointResponse(WriteCheckpointData("1000")) })
 
             syncStream =
                 SyncStream(
@@ -220,6 +228,7 @@ class SyncStreamTest {
                     retryDelayMs = 10,
                     logger = logger,
                     params = JsonObject(emptyMap()),
+                    scope = this,
                 )
 
             val job = launch { syncStream.streamingSync() }